Converted Login and SignUp pages to Class based components
