+++ frontend/src/App.js
@@ -2,8 +2,8 @@ import React, { Component } from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 
 import Home from "./components/pages/home/Home.js";
-import Login from "./components/pages/login/Login.js";
-import SignUp from "./components/pages/login/SignUp.js";
+import Login from "./components/layout/login/Login.js";
+import SignUp from "./components/layout/signup/SignUp.js";
 import StudentProfile from "./components/pages/student/Profile.js";
 
 import TeacherProfile from "./components/layout/teacher/Profile.js";
